+++ includes/frontend/class-header.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.4.0
+ * @version  2.4.1
  *
  * Contents:
  *
@@ -328,7 +328,7 @@ class Modern_Header {
 		 * HTML Body classes
 		 *
 		 * @since    1.0.0
-		 * @version  2.2.0
+		 * @version  2.4.1
 		 *
 		 * @param  array $classes
 		 */
@@ -379,12 +379,6 @@ class Modern_Header {
 						$classes[] = 'is-singular';
 
 						$post_id = get_the_ID();
-
-						// Privacy Policy page
-
-							if ( (int) get_option( 'wp_page_for_privacy_policy' ) === $post_id ) {
-								$classes[] = 'page-privacy-policy';
-							}
 
 						// Has featured image?